Add support for overlayfs
This introduces an option to mount the user's specified base directory
in overlayfs before setting up the filesystem namespace, allowing
an the base to be read-only inside the container with a writable
layer over top of it. File changes made inside the container will
be discarded when the container exits.